gpio: Introduce GPIO names
This introduces GPIO names. So far we only have labels which are given
by the requester. In contrast names are given by the provider and do not
change depending on whoever requests a GPIO. The gpio commands now also
accept to reference a GPIO by name.
The printing of the gpioinfo command is adjusted to nicely print both
the label and the name.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>