pstore: fix build by re-resolving botched merge conflict
297ac7288 ("pstore: Only capture log messages") removed the pstore console.
v3 of the eventually merged 75a74b5c ("console: fix out-of-bounds read
in dputc(/dev/*, ...)") reinstated part of this code, which broke the
pstore build because the now removed bufsize member had a user
reintroduced in the patch.

Fix this by reverting the offending patch's changes to the pstore code.
It wasn't applicable to begin with.

Fixes: 75a74b5c ("console: fix out-of-bounds read in dputc(/dev/*, ...)")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>