Revert "mci: core: make mci_register -EPROBE_DEFER sensitive"
The patch does the right thing: When a regulator can't be found and
-EPROBE_DEFER is returned we forward it. However, we have several boards
in the tree which do not have any driver for the regulator, but still
work because the regulator is enabled by default or by board code.

Revert this patch until a solution for these board is found.

This reverts commit 2e16c52ea277dd002235835a10f349233c31d965.