regulator: fixed: take regulator-boot-on into account
There are regulators which are turned on by default (hw-design) and
shouldn't be disabled during the boot. Upstream uses the
regulator-boot-on dt-property for it:

8<--------------------------------------------------
  regulator-boot-on:
    description: bootloader/firmware enabled regulator
    type: boolean
8<--------------------------------------------------

The difference between boot-on and always-on is the linux suspend
handling. Regulators marked as boot-on can be disabled during suspend.
This case isn't interesting for barebox so we can mark those as
always-on too.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>