mtd: nand-imx: Create BBT automatically when necessary
Due to the differences of the logical page format and the raw page
format on NAND the generic nand support can't read the bad block
marker on the NAND. For this reason we have the imx_nand_bbm command
which knows about these specialities and creates a BBT. The problem
with this command is that one has to call it and experience shows
this is often forgotten. Linux will then create a BBT based on wrong
informations and the real bad blocks may be lost.

With this patch we automatically create a BBT when none is found and
make the command unnecssary.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>