usb: gadget: fsl_udc: Add PBL image loading support
For boards that do the RAM setup in code we can up to now only download
the PBL part to SRAM. This patch adds support for downloading the rest
of the image after the RAM has been configured.

The ROM is nice enough to leave the USB controller initialized after a
download, so we don't have to reinitialize it, but can simply continue to
use the controller. Like all two-staged loading processes on i.MX this
needs board support, it will only work when a board calls
imx_barebox_load_usb() or one of the SoC specific variants.

This needs the host counterpart in imx-usb-loader which is done in the
next patch.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>