scripts: imx-usb-loader: download address is always the header address
We always have to upload the image beginning with the header, we have
to skip everything that is before the header.
In MODE_BULK this is explicit, but in HID mode this is obfuscated by
doing several calculations with helper variables. Drop this and simplify
code.

Tested on i.MX51, i.MX53 and i.MX6

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>